it will work

+++ models/movies.js
@@ -3,7 +3,7 @@ module.exports = (sequelize, DataTypes) => {
   const movies = sequelize.define('movies', {
     title: DataTypes.TEXT,
     year: DataTypes.INTEGER,
-    socialRating: DataTypes.DOUBLE,
+    socialRating: DataTypes.Numeric,
     movieRating: DataTypes.CHAR
   }, {});
   movies.associate = function(models) {
